Fix when line number is greater than last line
When I have a file

    wc -l file.test
    4

and open it with

   nvim file.test:5

the plugin throws an error.

The behavior of the plugin has been changed to only take the given line
number if it is smaler or equal to the number of lines in the file.
So running `nvim file.test:5` with this patch sets the cursor to line 4.

Co-authored-by: Lewis Russell <me@lewisr.dev>